refactor(agent): isolate transcript projection
