+++ packages/agent/src/presentation-session.ts
@@ -8,7 +8,6 @@ import type {
   PresentationCommand,
   PresentationDisplayState,
   PresentationSession,
-  PresentationTransientState,
   RepositoryInstructionsDisplay,
   SessionNaming,
   SessionSummary,
@@ -45,6 +44,12 @@ import {
   projectToolDisplays,
   resolveActionableChangePreviewReference,
 } from "./presentation-tool-projection.js";
+import {
+  projectActiveReasoningSnapshot,
+  projectTranscript,
+  providerDisplayName,
+  reasoningDisplayId,
+} from "./presentation-transcript-projection.js";
 import { listProjectPaths } from "./project-path-catalog.js";
 import {
   type SessionNamingHistoryState,
@@ -231,8 +236,8 @@ export async function createPresentationSession(
     }
     let transcript: readonly TranscriptItem[] = projectTranscript(
       records,
-      projectedOperations,
-      changePreviewCache,
+      projectedOperations.map(({ display }) => display),
+      projectToolDisplays(records, changePreviewCache),
     );
     const historyPageSize = boundedHistoryPageSize(options[presentationHistoryPageSize]);
     let loadedTranscriptStart = Math.max(0, transcript.length - historyPageSize);
@@ -799,8 +804,8 @@ export async function createPresentationSession(
       }
       const activatedTranscript = projectTranscript(
         activatedRecords,
-        activatedOperations,
-        changePreviewCache,
+        activatedOperations.map(({ display }) => display),
+        projectToolDisplays(activatedRecords, changePreviewCache),
       );
       const activatedLoadedTranscriptStart = Math.max(
         0,
@@ -1176,8 +1181,8 @@ export async function createPresentationSession(
             }
             transcript = projectTranscript(
               refreshedRecords,
-              refreshedOperations,
-              changePreviewCache,
+              refreshedOperations.map(({ display }) => display),
+              projectToolDisplays(refreshedRecords, changePreviewCache),
             );
             loadedTranscriptStart = Math.min(
               loadedTranscriptStart,
@@ -2914,393 +2919,6 @@ async function readPresentationSessionRecords(
   );
 }
 
-function projectTranscript(
-  records: readonly SourcedSessionRecord[],
-  operations: readonly ProjectedOperation[],
-  changePreviewCache: ReadonlyMap<string, ToolPreviewDisplay | null>,
-): readonly TranscriptItem[] {
-  const toolDisplays = projectToolDisplays(records, changePreviewCache);
-  const attemptProviders = new Map<string, string>(
-    records.flatMap(({ entry, sessionId }) =>
-      entry.schemaVersion === 3 && entry.record.type === "provider_attempt_started"
-        ? [
-            [
-              `${sessionId}:${entry.record.runId}:${entry.record.turn}:${entry.record.attempt}`,
-              providerDisplayName(entry.record.targetIdentity.vendor),
-            ] as const,
-          ]
-        : [],
-    ),
-  );
-  const reasoningStarts = new Map(
-    records.flatMap(({ entry, sessionId }) =>
-      entry.schemaVersion === 3 &&
-      entry.record.type === "runtime_event" &&
-      entry.record.event.type === "model_reasoning_started"
-        ? [
-            [
-              `${sessionId}:${entry.record.runId}:${entry.record.event.id}`,
-              entry.record.event,
-            ] as const,
-          ]
-        : [],
-    ),
-  );
-  const terminalBoundaries = new Map(
-    records.flatMap(({ entry, sessionId }) => {
-      if (entry.schemaVersion !== 3) {
-        return [];
-      }
-      if (entry.record.type === "runtime_event" && entry.record.event.type === "session_settled") {
-        return [
-          [`${sessionId}:${entry.record.runId}`, { sessionId, sequence: entry.sequence }] as const,
-        ];
-      }
-      return entry.record.type === "run_settled"
-        ? [[`${sessionId}:${entry.record.runId}`, { sessionId, sequence: entry.sequence }] as const]
-        : [];
-    }),
-  );
-  const publishedResponses = new Set(
-    records.flatMap(({ entry, sessionId }) =>
-      entry.schemaVersion === 3 && entry.record.type === "model_response_published"
-        ? [`${sessionId}:${entry.record.responseSequence}`]
-        : [],
-    ),
-  );
-  const completedInlineRuns = new Set(
-    records.flatMap(({ entry, sessionId }) =>
-      entry.schemaVersion === 3 &&
-      entry.record.type === "runtime_event" &&
-      entry.record.event.type === "model_message_completed"
-        ? [`${sessionId}:${entry.record.runId}`]
-        : [],
-    ),
-  );
-  const items: TranscriptItem[] = [];
-  const terminalNoticeRuns = new Set<string>();
-  for (const { entry, sessionId } of records) {
-    if (entry.schemaVersion !== 3) {
-      continue;
-    }
-    if (entry.record.type === "logical_run_started") {
-      items.push({
-        type: "user_message",
-        id: `${sessionId}:${entry.sequence}`,
-        sequence: entry.sequence,
-        sourceSessionId: sessionId,
-        branchBoundary: null,
-        text: entry.record.userMessage,
-      });
-      continue;
-    }
-    if (entry.record.type === "runtime_event" && entry.record.event.type === "tool_requested") {
-      const tool = toolDisplays.get(`${sessionId}:${entry.record.event.callId}`);
-      if (tool !== undefined) {
-        items.push(tool);
-      }
-      continue;
-    }
-    if (entry.record.type === "context_compaction_committed") {
-      items.push({
-        type: "compaction_marker",
-        id: `${sessionId}:${entry.sequence}`,
-        sequence: entry.sequence,
-        sourceSessionId: sessionId,
-        branchBoundary: null,
-        windowNumber: entry.record.windowNumber,
-        sourceThrough: entry.record.sourceThrough,
-        retainedFrom: entry.record.retainedFrom,
-      });
-      continue;
-    }
-    if (
-      entry.record.type === "runtime_event" &&
-      entry.record.event.type === "model_reasoning_settled" &&
-      entry.record.event.status !== "completed"
-    ) {
-      const start = reasoningStarts.get(
-        `${sessionId}:${entry.record.runId}:${entry.record.event.id}`,
-      );
-      const attemptIdentity = /^(\d+):(\d+):/.exec(entry.record.event.id);
-      if (start !== undefined && attemptIdentity !== null) {
-        items.push({
-          type: "reasoning_block",
-          id: reasoningDisplayId(sessionId, entry.record.runId, entry.record.event.id),
-          sequence: entry.sequence,
-          sourceSessionId: sessionId,
-          branchBoundary: terminalBoundaries.get(`${sessionId}:${entry.record.runId}`) ?? null,
-          artifactType: start.artifactType,
-          disclosure: "owner_only",
-          provider:
-            attemptProviders.get(
-              `${sessionId}:${entry.record.runId}:${attemptIdentity[1]}:${attemptIdentity[2]}`,
-            ) ?? providerDisplayName(undefined),
-          status: entry.record.event.status,
-          text: null,
-          artifact: null,
-        });
-      }
-      continue;
-    }
-    if (
-      entry.record.type === "runtime_event" &&
-      entry.record.event.type === "session_interrupted"
-    ) {
-      items.push({
-        type: "session_notice",
-        id: `${sessionId}:${entry.sequence}`,
-        sequence: entry.sequence,
-        sourceSessionId: sessionId,
-        branchBoundary: terminalBoundaries.get(`${sessionId}:${entry.record.runId}`) ?? null,
-        status: "interrupted",
-        reason: entry.record.event.reason,
-      });
-      continue;
-    }
-    if (
-      entry.record.type === "provider_attempt_interrupted" &&
-      entry.record.reason === "process_restart"
-    ) {
-      items.push({
-        type: "session_notice",
-        id: `${sessionId}:${entry.sequence}`,
-        sequence: entry.sequence,
-        sourceSessionId: sessionId,
-        branchBoundary: null,
-        status: "interrupted",
-        reason: "process_restart",
-      });
-      continue;
-    }
-    if (
-      entry.record.type === "runtime_event" &&
-      entry.record.event.type === "session_settled" &&
-      (entry.record.event.result.status === "failed" ||
-        entry.record.event.result.status === "incomplete") &&
-      !terminalNoticeRuns.has(`${sessionId}:${entry.record.runId}`)
-    ) {
-      terminalNoticeRuns.add(`${sessionId}:${entry.record.runId}`);
-      const result = entry.record.event.result;
-      items.push(
-        result.status === "failed"
-          ? {
-              type: "session_notice",
-              id: `${sessionId}:${entry.sequence}`,
-              sequence: entry.sequence,
-              sourceSessionId: sessionId,
-              branchBoundary: { sessionId, sequence: entry.sequence },
-              status: "failed",
-              code: result.error.code,
-              message: safeRunFailureMessage(result.error.code),
-            }
-          : {
-              type: "session_notice",
-              id: `${sessionId}:${entry.sequence}`,
-              sequence: entry.sequence,
-              sourceSessionId: sessionId,
-              branchBoundary: { sessionId, sequence: entry.sequence },
-              status: "incomplete",
-              reason: result.reason,
-            },
-      );
-      continue;
-    }
-    if (
-      entry.record.type === "run_settled" &&
-      entry.record.status === "incomplete" &&
-      !terminalNoticeRuns.has(`${sessionId}:${entry.record.runId}`)
-    ) {
-      terminalNoticeRuns.add(`${sessionId}:${entry.record.runId}`);
-      items.push({
-        type: "session_notice",
-        id: `${sessionId}:${entry.sequence}`,
-        sequence: entry.sequence,
-        sourceSessionId: sessionId,
-        branchBoundary: { sessionId, sequence: entry.sequence },
-        status: "incomplete",
-        reason: entry.record.reason,
-      });
-      continue;
-    }
-    if (entry.record.type !== "model_response_completed") {
-      continue;
-    }
-    const artifactBacked =
-      entry.record.response.recordVersion === 2 &&
-      (entry.record.response.text.storage === "artifact" ||
-        entry.record.response.reasoning?.storage === "artifact");
-    if (
-      (artifactBacked && !publishedResponses.has(`${sessionId}:${entry.sequence}`)) ||
-      (!artifactBacked && !completedInlineRuns.has(`${sessionId}:${entry.record.runId}`))
-    ) {
-      continue;
-    }
-    const reasoningField = entry.record.response.reasoning;
-    if (reasoningField !== undefined) {
-      const reasoningText =
-        typeof reasoningField === "string"
-          ? reasoningField
-          : reasoningField.storage === "inline"
-            ? reasoningField.text
-            : null;
-      const reasoningArtifact =
-        typeof reasoningField !== "string" && reasoningField.storage === "artifact"
-          ? {
-              id: reasoningField.reference.id,
-              mediaType: reasoningField.reference.mediaType,
-              byteCount: reasoningField.reference.byteCount,
-              source: "model_response" as const,
-            }
-          : null;
-      if (reasoningArtifact !== null || (reasoningText !== null && reasoningText.length > 0)) {
-        items.push({
-          type: "reasoning_block",
-          id: reasoningDisplayId(
-            sessionId,
-            entry.record.runId,
-            `${entry.record.turn}:${entry.record.attempt}:provider-reasoning-0`,
-          ),
-          sequence: entry.sequence,
-          sourceSessionId: sessionId,
-          branchBoundary: terminalBoundaries.get(`${sessionId}:${entry.record.runId}`) ?? null,
-          artifactType: "provider_reasoning",
-          disclosure: "owner_only",
-          provider: providerDisplayName(entry.record.targetIdentity.vendor),
-          status: "completed",
-          text: reasoningText,
-          artifact: reasoningArtifact,
-        });
-      }
-    }
-    const text =
-      entry.record.response.recordVersion === 2
-        ? entry.record.response.text.storage === "inline"
-          ? entry.record.response.text.text
-          : null
-        : entry.record.response.text;
-    const artifact =
-      entry.record.response.recordVersion === 2 && entry.record.response.text.storage === "artifact"
-        ? {
-            id: entry.record.response.text.reference.id,
-            mediaType: entry.record.response.text.reference.mediaType,
-            byteCount: entry.record.response.text.reference.byteCount,
-            source: "model_response" as const,
-          }
-        : null;
-    if (artifact === null && (text === null || text.length === 0)) {
-      continue;
-    }
-    items.push({
-      type: "assistant_message",
-      id: `${sessionId}:${entry.sequence}`,
-      sequence: entry.sequence,
-      sourceSessionId: sessionId,
-      branchBoundary: terminalBoundaries.get(`${sessionId}:${entry.record.runId}`) ?? null,
-      text,
-      artifact,
-    });
-  }
-  const recordOrder = new Map(
-    records.map(
-      (record, index) => [`${record.sessionId}:${record.entry.sequence}`, index] as const,
-    ),
-  );
-  const itemOrder = new Map(items.map((item, index) => [item, index] as const));
-  const operationLinks: TranscriptItem[] = operations.map(({ display }) => ({
-    type: "operation_link",
-    id: `operation:${display.operationId}`,
-    operationId: display.operationId,
-    sequence: display.origin.sourceSequence,
-    sourceSessionId: display.origin.sessionId,
-    branchBoundary: {
-      sessionId: display.origin.sessionId,
-      sequence: display.origin.sourceSequence,
-    },
-  }));
-  return [...items, ...operationLinks].sort((left, right) => {
-    const leftRecord = recordOrder.get(`${left.sourceSessionId}:${left.sequence}`) ?? Infinity;
-    const rightRecord = recordOrder.get(`${right.sourceSessionId}:${right.sequence}`) ?? Infinity;
-    if (leftRecord !== rightRecord) {
-      return leftRecord - rightRecord;
-    }
-    if (left.type === "operation_link" && right.type === "operation_link") {
-      return left.operationId.localeCompare(right.operationId);
-    }
-    if (left.type === "operation_link") {
-      return 1;
-    }
-    if (right.type === "operation_link") {
-      return -1;
-    }
-    return (itemOrder.get(left) ?? 0) - (itemOrder.get(right) ?? 0);
-  });
-}
-
-function projectActiveReasoningSnapshot(input: {
-  readonly records: readonly SourcedSessionRecord[];
-  readonly sessionId: string;
-  readonly runId: string;
-  readonly expectedId: string;
-  readonly event: Extract<RuntimeEvent, { readonly type: "model_reasoning_updated" }>;
-  readonly afterSequence: number;
-  readonly provider: string;
-}): NonNullable<PresentationTransientState["reasoning"]> | undefined {
-  let start: Extract<RuntimeEvent, { readonly type: "model_reasoning_started" }> | undefined;
-  let startSequence = 0;
-  for (const { entry, sessionId } of input.records) {
-    if (
-      sessionId !== input.sessionId ||
-      entry.schemaVersion !== 3 ||
-      entry.record.type !== "runtime_event" ||
-      entry.record.runId !== input.runId
-    ) {
-      continue;
-    }
-    const event = entry.record.event;
-    if (
-      event.type === "model_reasoning_started" &&
-      reasoningDisplayId(sessionId, input.runId, event.id) === input.expectedId
-    ) {
-      start = event;
-      startSequence = entry.sequence;
-      continue;
-    }
-    if (
-      start !== undefined &&
-      entry.sequence > startSequence &&
-      event.type === "model_reasoning_settled" &&
-      event.id === start.id
-    ) {
-      return undefined;
-    }
-  }
-  if (start === undefined || start.id !== input.event.id) {
-    return undefined;
-  }
-  return {
-    id: input.expectedId,
-    afterSequence: input.afterSequence,
-    artifactType: start.artifactType,
-    disclosure: "owner_only",
-    provider: input.provider,
-    status: "active",
-    text: input.event.text,
-  };
-}
-
-function reasoningDisplayId(
-  sessionId: string | null,
-  runId: string | null,
-  runtimeReasoningId: string,
-): string {
-  return `${sessionId ?? "unknown-session"}:${runId ?? "unknown-run"}:${runtimeReasoningId}`;
-}
-
-function providerDisplayName(vendor: string | undefined): string {
-  return vendor === "deepseek" ? "DeepSeek" : (vendor ?? "Provider");
-}
-
 async function projectPendingInteractions(
   records: readonly SourcedSessionRecord[],
   options: Pick<CreatePresentationSessionOptions, "stateRoot" | "workspaceRoot">,
@@ -3417,22 +3035,6 @@ async function projectChangePreview(
   } catch {
     return null;
   }
-}
-
-function safeRunFailureMessage(code: string): string {
-  if (code === "tool_effect_indeterminate") {
-    return "A tool effect requires inspection before continuing.";
-  }
-  if (code === "session_persistence_failed") {
-    return "The session could not make its result durable.";
-  }
-  if (code.startsWith("context_") || code.startsWith("token_")) {
-    return "The run could not continue within its context limits.";
-  }
-  if (code === "skill_activation_failed") {
-    return "The requested Skill activation failed.";
-  }
-  return "The model run failed.";
 }
 
 function projectRepositoryInstructions(
